+++ test/e2e/scenarios/balancing/node3.js
@@ -1,0 +1,15 @@
+const { createNode, logEventEmitting } = require("../../utils");
+
+const broker = createNode("node3");
+broker.loadService("./test.service.js");
+
+broker.createService({
+	name: "payment",
+	events: {
+		"user.created"(ctx) {
+			logEventEmitting(this, ctx);
+		}
+	}
+});
+
+broker.start();